services/desktop_app/requirements.txt
PySide6
httpx
python-dotenv